# backend/docs/server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for CI/CD application developed using Go (Gin, Gorm).",
        "title": "CI/CD Application - Server API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/pipelines": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipelines"
                ],
                "summary": "Get all pipelines",
                "operationId": "all-pipelines",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "description": "Filter",
                        "name": "pipeline",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.PipelineCore"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of pipelines",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.PipelineShort"
                            }
                        }
                    },
                    "400": {
                        "description": "Error in request",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No records found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipelines"
                ],
                "summary": "Create new pipeline",
                "operationId": "create-pipeline",
                "parameters": [
                    {
                        "description": "New pipeline entry",
                        "name": "pipeline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PipelineCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Newly created pipeline",
                        "schema": {
                            "$ref": "#/definitions/model.Pipeline"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
        "/pipelines/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipelines"
                ],
                "summary": "Get the single pipeline",
                "operationId": "single-pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Requested pipeline",
                        "schema": {
                            "$ref": "#/definitions/model.Pipeline"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No record found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipelines"
                ],
                "summary": "Update pipeline",
                "operationId": "update-pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated pipeline entry",
                        "name": "pipeline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PipelineCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated pipeline",
                        "schema": {
                            "$ref": "#/definitions/model.Pipeline"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipelines"
                ],
                "summary": "Delete pipeline",
                "operationId": "delete-pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete message",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Build": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildStep"
                    }
                },
                "tags": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
...
        "model.BuildLog": {
            "type": "object",
            "properties": {
                "build_step_id": {
                    "type": "string"
                },
                "command": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                }
            }
        },
...
        "model.BuildStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "BuildRunning",
                "BuildSuccessful",
                "BuildFailed",
                "BuildCanceled"
            ]
        },
        "model.BuildStep": {
            "type": "object",
            "properties": {
                "build_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "id": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildLog"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Pipeline": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "builds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Build"
                    }
                },
                "config": {
                    "$ref": "#/definitions/model.PipelineConfig"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "secrets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Secret"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "variables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Variable"
                    }
                }
            }
        },
        "model.PipelineConfig": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "commands": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "name": {
                                "type": "string"
                            }
                        }
                    }
                },
                "system": {
                    "type": "string"
                }
            }
        },
        "model.PipelineCore": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "config": {
                    "$ref": "#/definitions/model.PipelineConfig"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                }
            }
        },
        "model.PipelineCreate": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "config": {
                    "$ref": "#/definitions/model.PipelineConfig"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                }
            }
        },
        "model.PipelineShort": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "config": {
                    "$ref": "#/definitions/model.PipelineConfig"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Secret": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Variable": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "time.Duration": {
            "type": "integer",
            "enum": [
                1,
                1000,
                1000000,
                1000000000
            ],
            "x-enum-varnames": [
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second"
            ]
        },
        "util.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}